Update camera-view.md
Some of the code examples in the documentation for CameraView are missing the x:Name property in the XAML. The name property is referenced by the button command property Command="{Binding StartCameraPreviewCommand, Source={x:Reference Camera}}”.  Also, I think it is important to note that the CameraView is in a separate Nuget Package from the CommunityToolkit.Maui. The getting started section link is fine but no one will click it if they believe the CameraView is in the CommunityToolkit.Maui package.